fix: Don't enhance H2s if theres a Table Of Contents
