+++ dotcom-rendering/src/model/enhance-interactive-contents-elements.ts
@@ -1,23 +1,10 @@
+import { isLegacyTableOfContents } from './isLegacyTableOfContents';
 import { stripHTML } from './sanitise';
-
-const scriptUrls = [
-	'https://interactive.guim.co.uk/page-enhancers/nav/boot.js',
-	'https://uploads.guim.co.uk/2019/03/20/boot.js',
-	'https://uploads.guim.co.uk/2019/12/11/boot.js',
-	'https://interactive.guim.co.uk/testing/2020/11/voterSlideshow/boot.js',
-	'https://uploads.guim.co.uk/2021/10/15/boot.js',
-];
-
-const isInteractiveContentsBlockElement = (element: CAPIElement): boolean =>
-	element._type ===
-		'model.dotcomrendering.pageElements.InteractiveBlockElement' &&
-	!!element.scriptUrl &&
-	scriptUrls.includes(element.scriptUrl);
 
 const enhance = (elements: CAPIElement[]): CAPIElement[] => {
 	const updatedElements: CAPIElement[] = [];
 	const hasInteractiveContentsBlockElement = elements.some((element) =>
-		isInteractiveContentsBlockElement(element),
+		isLegacyTableOfContents(element),
 	);
 
 	if (hasInteractiveContentsBlockElement) {
@@ -40,7 +27,7 @@ const enhance = (elements: CAPIElement[]): CAPIElement[] => {
 
 		// replace interactive content block
 		elements.forEach((element) => {
-			if (isInteractiveContentsBlockElement(element)) {
+			if (isLegacyTableOfContents(element)) {
 				updatedElements.push({
 					_type: 'model.dotcomrendering.pageElements.DividerBlockElement',
 					size: 'full',
